requirements.txt
opencv~=4.5.3
numpy~=1.21.2
pillow~=8.2.0
pyqt~=5.9.2
python~=3.8
